requirements.txt
streamlit
pandas
numpy
scikit-learn
yfinance
plotly
textblob
newsapi-python
matplotlib
reportlab
